backend/pyproject.toml
[project]
name = "viral-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi[standard]>=0.136.0",
  "google-genai>=1.70.0",
  "numpy>=2.4.3",
  "ocrmac>=1.0.1",
  "openai>=2.30.0",
  "opencv-python-headless>=4.13.0.92",
  "paddleocr>=3.4.0",
  "python-dotenv>=1.2.2",
  "send2trash>=2.1.0",
]

[tool.fastapi]
entrypoint = "src.main:app"
